+++ src/main/kotlin/com/chattriggers/ctjs/utils/config/ConfigString.kt
@@ -59,10 +59,10 @@ class ConfigString
 
         val middle = Renderer.screen.getWidth() / 2
 
-        Rectangle(-0x80000000, (middle - 105 + this.x).toFloat(), (this.y - 5).toFloat(), 210f, 45f)
-                .setShadow(-0x30000000, 3f, 3f)
+        Rectangle(-0x80000000, (middle - 105 + this.x).toDouble(), (this.y - 5).toDouble(), 210.0, 45.0)
+                .setShadow(-0x30000000, 3.0, 3.0)
                 .draw()
-        Text(this.name!!, (middle - 100 + this.x).toFloat(), this.y.toFloat()).draw()
+        Text(this.name!!, (middle - 100 + this.x).toDouble(), this.y.toDouble()).draw()
 
         //#if MC<=10809
         this.textField!!.xPosition = middle - 100 + this.x
